Move feature flag classes into own files
